requirements.txt
streamlit
tf-keras
pillow
numpy
opencv-python